requirements.txt
eventlet
python-socketio
flask
requests
dumper